refactor(code_gen): extract fs builtins into builtins/fs.rs
